fixed different yaw/pitch speed in ui on non-4:3 screens

+++ codemp/ui/ui_shared.h
@@ -449,8 +449,8 @@ typedef struct {
   float			bias;
   int				realTime;
   int				frameTime;
-	int				cursorx;
-	int				cursory;
+	float			cursorx;
+	float			cursory;
 	qboolean	debug;
 
   cachedAssets_t Assets;
